Displaying the Maxmind and Browscap version numbers under the Maintenance tab, to better troubleshoot issues related to these files

+++ admin/config/maintenance.php
@@ -276,33 +276,35 @@ $slim_browsers_exists =wp_slimstat::$wpdb->get_col( "SHOW TABLES LIKE '{$GLOBALS
 	</tr>
 	<tr>
 		<th scope="row">
-			<?php if (!file_exists(wp_slimstat::$maxmind_path)): ?>
-			<a class="button-secondary" href="<?php echo wp_slimstat_admin::$config_url.$current_tab ?>&amp;action=download-maxmind"
-				onclick="return(confirm('<?php _e('Do you want to download and install the geolocation database from MaxMind\'s server?','wp-slimstat'); ?>'))"><?php _e("Install GeoLite DB",'wp-slimstat'); ?></a>
-			<?php else: ?>
-			<a class="button-secondary" href="<?php echo wp_slimstat_admin::$config_url.$current_tab ?>&amp;action=delete-maxmind"
-				onclick="return(confirm('<?php _e('Do you want to uninstall the geolocation database?','wp-slimstat'); ?>'))"><?php _e("Uninstall GeoLite DB",'wp-slimstat'); ?></a>
-			<?php endif; ?>
+		<?php if ( !file_exists( wp_slimstat::$maxmind_path ) ): ?>
+			<a class="button-secondary" href="<?php echo wp_slimstat_admin::$config_url.$current_tab ?>&amp;action=download-maxmind" onclick="return( confirm( '<?php _e( 'Do you want to download and install the geolocation database from MaxMind\'s server?', 'wp-slimstat' ); ?>' ) )"><?php _e( 'Install GeoLite DB', 'wp-slimstat' ); ?></a>
+		<?php else: ?>
+			<a class="button-secondary" href="<?php echo wp_slimstat_admin::$config_url.$current_tab ?>&amp;action=delete-maxmind" onclick="return( confirm( '<?php _e( 'Do you want to uninstall the geolocation database?', 'wp-slimstat' ); ?>' ) )"><?php _e( 'Uninstall GeoLite DB', 'wp-slimstat' ); ?></a>
+		<?php endif; ?>
 		</th>
-		<td>
-			<span class="description"><?php _e("The <a href='https://dev.maxmind.com/geoip/geoip2/geolite2/' target='_blank'>MaxMind GeoLite2 library</a>, which Slimstat uses to geolocate visitors, is released under the Creative Commons BY-SA 4.0 license, and cannot be directly bundled with the plugin because of license incompatibility issues. We are mandated to have the user take an affirmative action in order to enable this functionality. If you're experiencing issues, please <a href='https://slimstat.freshdesk.com/solution/articles/12000039798-how-to-manually-install-the-maxmind-geolocation-data-file-' target='_blank'>take a look at our knowledge base</a> to learn how to install this file manually.", 'wp-slimstat' ) ?></span>
+		<td><?php
+			$maxmind_last_modified = '';
+			if ( file_exists( wp_slimstat::$maxmind_path ) && false !== ( $file_stat = @stat( wp_slimstat::$maxmind_path ) ) ) { 
+				$maxmind_last_modified = date_i18n( wp_slimstat::$settings[ 'date_format' ], $file_stat[ 'mtime' ] );
+			} 
+			?>
+			<span class="description"><?php _e("The <a href='https://dev.maxmind.com/geoip/geoip2/geolite2/' target='_blank'>MaxMind GeoLite2 library</a>, which Slimstat uses to geolocate visitors, is released under the Creative Commons BY-SA 4.0 license, and cannot be directly bundled with the plugin because of license incompatibility issues. We are mandated to have the user take an affirmative action in order to enable this functionality. If you're experiencing issues, please <a href='https://slimstat.freshdesk.com/solution/articles/12000039798-how-to-manually-install-the-maxmind-geolocation-data-file-' target='_blank'>take a look at our knowledge base</a> to learn how to install this file manually.", 'wp-slimstat' ); if ( !empty( $maxmind_last_modified ) ) { echo ' ' . __( 'Your data file was last downloaded on', 'wp-slimstat' ) . ' <strong>' . $maxmind_last_modified . '</strong>.'; } ?></span>
 		</td>
 	</tr>
-	<tr class="alternate">
-		<th scope="row">
-			<?php if ( !file_exists( slim_browser::$browscap_autoload_path ) ) : ?>
-			<a class="button-secondary" href="<?php echo wp_slimstat_admin::$config_url.$current_tab ?>&amp;action=download-browscap"
-				onclick="return( confirm( '<?php _e( 'Do you want to download and install the Browscap data file from our server?', 'wp-slimstat' ); ?>' ) )"><?php _e( 'Install Browscap', 'wp-slimstat' ); ?></a>
-				<br><br>&nbsp;<strong><?php _e( 'Requires PHP 7.1', 'wp-slimstat' ) ?></strong>
-			<?php else: ?>
-			<a class="button-secondary" href="<?php echo wp_slimstat_admin::$config_url.$current_tab ?>&amp;action=delete-browscap"
-				onclick="return( confirm( '<?php _e( 'Do you want to uninstall the Browscap data file?', 'wp-slimstat' ); ?>' ) )"><?php _e( 'Uninstall Browscap', 'wp-slimstat' ); ?></a>
-			<?php endif; ?>
-		</th>
-		<td>
-			<span class="description"><?php _e( "We are contributing to the <a href='https://browscap.org/' target='_blank'>Browscap Capabilities Project</a>, which we use to decode your visitors' user agent string into browser name and operating system. We use an <a href='https://github.com/slimstat/browscap-db' target='_blank'>optimized version of their data structure</a>, for improved performance. Slimstat can use this data file instead of the built-in heuristic function, to accurately determine your visitors' browser information. It also checks for updates and downloads the latest version for you. Do not hesitate to <a href='https://support.wp-slimstat.com' target='_blank'>contact our support team</a> if you have any questions.", 'wp-slimstat' ) ?></span>
-		</td>
-	</tr>
+	<?php 
+		if ( version_compare( PHP_VERSION, '7.1', '>=' ) ) {
+			echo '<tr class="alternate"><th scope="row">';
+			$browscap_version = '';
+			if ( !file_exists( slim_browser::$browscap_autoload_path ) ) {
+				echo '<a class="button-secondary" href="' . wp_slimstat_admin::$config_url.$current_tab . '&amp;action=download-browscap">' . __( 'Install Browscap', 'wp-slimstat' ) . '</a>';
+			}
+			else {
+				echo '<a class="button-secondary" href="' . wp_slimstat_admin::$config_url.$current_tab . '&amp;action=delete-browscap">' . __( 'Uninstall Browscap', 'wp-slimstat' ) . '</a>';
+				$browscap_version = ' ' . sprintf( __( 'You are currently using version %s.' ), '<strong>' . slim_browser::$browscap_local_version . '</strong>');
+			}
+			echo '</th><td><span class="description">' . __( "We are contributing to the <a href='https://browscap.org/' target='_blank'>Browscap Capabilities Project</a>, which we use to decode your visitors' user agent string into browser name and operating system. We use an <a href='https://github.com/slimstat/browscap-db' target='_blank'>optimized version of their data structure</a>, for improved performance. Slimstat can use this data file instead of the built-in heuristic function, to accurately determine your visitors' browser information. It also checks for updates and downloads the latest version for you. Do not hesitate to <a href='https://support.wp-slimstat.com' target='_blank'>contact our support team</a> if you have any questions.", 'wp-slimstat' ) . $browscap_version . '</td></tr>';
+		}
+	?>
 	<tr>
 		<td colspan="2" class="slimstat-options-section-header" id="wp-slimstat-configuration-string"><?php _e('Configuration String','wp-slimstat') ?></td>
 	</tr>
